move DocsTable into Table
