--- packages/styleguide/src/components/DocsTable/index.js
@@ -1,1 +1,0 @@
-export { default } from './DocsTable';
