+++ packages/styleguide/src/components/index.js
@@ -6,7 +6,6 @@ export { default as Code, CodeBlock, CodeJSX, CodeBlockJSX } from './Code';
 export { default as ColorPalette } from './ColorPalette';
 export { default as ComponentDocs } from './ComponentDocs';
 export { default as ComponentInfo } from './ComponentInfo';
-export { default as DocsTable } from './DocsTable';
 export { default as Icon } from './Icon';
 export {
   default as Note,
@@ -18,5 +17,5 @@ export {
 } from './Note';
 export { default as Page } from './Page';
 export { default as Preview } from './Preview';
-export { default as Table } from './Table';
+export { default as Table, DocsTable } from './Table';
 export { H1, H2, H3, H4, H5, P, Link } from './Typography';
